Separate filters into different TUs
