+++ entity/BUILD.gn
@@ -37,8 +37,12 @@ impeller_component("entity") {
     "contents/content_context.h",
     "contents/contents.cc",
     "contents/contents.h",
-    "contents/filter_contents.cc",
-    "contents/filter_contents.h",
+    "contents/filters/blend_filter_contents.cc",
+    "contents/filters/blend_filter_contents.h",
+    "contents/filters/filter_contents.cc",
+    "contents/filters/filter_contents.h",
+    "contents/filters/gaussian_blur_filter_contents.cc",
+    "contents/filters/gaussian_blur_filter_contents.h",
     "contents/linear_gradient_contents.cc",
     "contents/linear_gradient_contents.h",
     "contents/solid_color_contents.cc",
