+++ entity/contents/filters/filter_contents.h
@@ -8,17 +8,12 @@
 #include <variant>
 #include <vector>
 
-#include "impeller/entity/contents/contents.h"
 #include "impeller/entity/entity.h"
 #include "impeller/renderer/formats.h"
 
 namespace impeller {
 
 class Pipeline;
-
-/*******************************************************************************
- ******* FilterContents
- ******************************************************************************/
 
 class FilterContents : public Contents {
  public:
@@ -83,68 +78,6 @@ class FilterContents : public Contents {
   Rect destination_;
 
   FML_DISALLOW_COPY_AND_ASSIGN(FilterContents);
-};
-
-/*******************************************************************************
- ******* BlendFilterContents
- ******************************************************************************/
-
-class BlendFilterContents : public FilterContents {
- public:
-  using AdvancedBlendProc = std::function<bool(
-      const std::vector<std::shared_ptr<Texture>>& input_textures,
-      const ContentContext& renderer,
-      RenderPass& pass)>;
-
-  BlendFilterContents();
-
-  ~BlendFilterContents() override;
-
-  void SetBlendMode(Entity::BlendMode blend_mode);
-
- private:
-  // |FilterContents|
-  bool RenderFilter(const std::vector<std::shared_ptr<Texture>>& input_textures,
-                    const ContentContext& renderer,
-                    RenderPass& pass) const override;
-
-  Entity::BlendMode blend_mode_;
-  AdvancedBlendProc advanced_blend_proc_;
-
-  FML_DISALLOW_COPY_AND_ASSIGN(BlendFilterContents);
-};
-
-/*******************************************************************************
- ******* DirectionalGaussionBlurFilterContents
- ******************************************************************************/
-
-class DirectionalGaussianBlurFilterContents final : public FilterContents {
- public:
-  DirectionalGaussianBlurFilterContents();
-
-  ~DirectionalGaussianBlurFilterContents() override;
-
-  void SetRadius(Scalar radius);
-
-  void SetDirection(Vector2 direction);
-
-  void SetClipBorder(bool clip);
-
- private:
-  // |FilterContents|
-  bool RenderFilter(const std::vector<std::shared_ptr<Texture>>& input_textures,
-                    const ContentContext& renderer,
-                    RenderPass& pass) const override;
-
-  // |FilterContents|
-  virtual ISize GetOutputSize(
-      const InputTextures& input_textures) const override;
-
-  Scalar radius_ = 0;
-  Vector2 direction_;
-  bool clip_ = false;
-
-  FML_DISALLOW_COPY_AND_ASSIGN(DirectionalGaussianBlurFilterContents);
 };
 
 }  // namespace impeller
